improve stereo by auto-setting params
Improves stereo rendering by adjusting the focal distance and eye
separation based on the focus point in rviz.

Without this change stereo rendering works ok, but parameters in the
view controller need to be tweaked to get good results.  Zooming in or
out req uires re-tweaking the parameters.

With this change these parameters are tweaked automatically based on
the rviz view controller focus point (lookAt point) and the size of
the window on the screen.  The screen resolution (dots per inch) can
still be tweaked by hand (default of 81dpi works on a 27 inch
1920x1080 monitor).  The auto-tweaking can be disabled with view a
controller property to revert to the old behavior of setting the focal
distance by hand.